+++ src/components/ui/label/LabelTheme.ts
@@ -1,7 +1,7 @@
 import type { BaseTypographyComponentTheme, DefaultSizedLayoutClassMappers } from "../theme/common";
 import type { BorderClassMapper, RingClassMapper, WrapClassMapper, DirectionClassMapper, CursorClassMapper, FlexClassMapper, ShrinkClassMapper } from "../theme/layout";
 import type { GapClassMapper, FontSizeClassMapper, LineHeightClassMapper } from "../theme/size";
-import type { SimpleConsumerClassMapper, ShadowAppearanceClassMapper } from "../theme/appearance";
+import type { SimpleConsumerClassMapper, ShadowAppearanceClassMapper, DisabledOpacityClassMapper } from "../theme/appearance";
 
 export interface LabelTheme extends BaseTypographyComponentTheme {
   size: {
@@ -23,5 +23,6 @@ export interface LabelTheme extends BaseTypographyComponentTheme {
     cursor: CursorClassMapper;
     flex: FlexClassMapper;
     shrink: ShrinkClassMapper;
+    disabled: DisabledOpacityClassMapper;
   };
 }